+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 -   Rename experimental_analyze_model_and_dataset to analyze_model_and_dataset
 -   Add new GBT loss function `POISSON` for Poisson log likelihood.
 -   Go API: Categorical string values available for inspection.
+-   Improved training speed for unit-weight datasets.
 
 ### Fix